Refactoring metrix db code.
I've added a database.py module which has a MetrixDB class which
provides a high-level programatic interface to the database. The methods
of this class are mirrored by the command line programs which can now be
called as metrix_db.${program}

I've refactored the old scripts into classes inside their own modules.